+++ src/pages/signin/SignInPageLayout/index.native.js
@@ -1,36 +1,3 @@
-import React from 'react';
-import {Image, Text, View} from 'react-native';
-import PropTypes from 'prop-types';
-import styles from '../../../styles/styles';
-import logo from '../../../../assets/images/expensify-logo-coin-2x.png';
+import SignInPageLayoutNarrow from './SignInPageLayoutNarrow';
 
-const propTypes = {
-    // The children to show inside the layout
-    children: PropTypes.node.isRequired,
-};
-
-const SignInPageLayout = ({children}) => (
-    <View style={[styles.signInPageInnerNative]}>
-        <View style={[styles.signInPageLogoNative]}>
-            <Image
-                resizeMode="contain"
-                style={[styles.signinLogo]}
-                source={logo}
-            />
-        </View>
-
-        <View style={[styles.mt5, styles.mb4]}>
-            <Text style={[styles.h1]}>
-                Expensify.cash
-            </Text>
-        </View>
-
-        {children}
-    </View>
-);
-
-SignInPageLayout.propTypes = propTypes;
-SignInPageLayout.displayName = 'SignInPageLayout';
-
-
-export default SignInPageLayout;
+export default SignInPageLayoutNarrow;
